requirements.txt
numpy
pandas
matplotlib
seaborn
gensim
google-generativeai
anthropic
python-dotenv
tqdm
jupyter
ollama